requirements.txt
delocate==0.10.4
packaging==23.1
Pillow==9.5.0
typing_extensions==4.6.3
